pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["GabrielTeodoroSan <gabrielteodorodomingues@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
httpx = "^0.27.0"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.13.0"
isort = "^5.13.2"
blue = "^0.9.1"
pytest = "^8.3.2"


[tool.taskipy.tasks]
run = 'uvicorn app.main:app --host 0.0.0.0 --reload'
test = 'export PYTHONPATH="$(pwd)/app:$PYTHONPATH"; pytest -cov=app'
clear = 'isort . && blue .'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
